requirements.txt
groq>=0.12.0
python-dotenv>=1.0.0
pandas>=2.2.0
pydantic>=2.6.0
streamlit>=1.39.0
pytest>=8.0.0
